virtio-gpu: Add virtiogpu implementation
virtiogpu: enable hotplug support

Some hardcoded display IDs were removed and if we change the display
layout via "VBoxManage modifyvm <vm-name> setscreenlayout" we pass the
required information to enable/disable monitors to the virtio gpu.

Further, the VBoxManage setscreenlayout call is extended to pass the
display information to the virtiogpu in case it is used as a graphics
adapter.

virtio-gpu: Disable x2apic msr hint conditionally

If the VirtioGPU is used as a graphics controller and the guest is
configured with >8 CPUs, we disable the X2APIC_MSR HyperV hint.

We do this because there are known issues with the DVServerKMD virtio
gpu driver in Windows 10, that prevent the Windows from booting if the
specific HyperV hint is set. Not setting the hint allows the Windows to
boot. Therefore, we only disable the hint in the specific combination of
VirtioGPU + >8 CPUs in order to not affect guest performance in all
other cases.

VBox: virtio: add PDMR3QueryDevice to VMM vTable again

VBox:Virtio: Move from SVP include directory to cyberus include directory

VBox:Virtio: rename SVP Display to CBS Display

virtio: remove slvm dependency

vga: re-add dummy connector attachment

We previously had some VBOX_WITH_SVP define around the code connecting
our dummy connector when switching from the VGA display to some custom
display (either GVT or VirtipGPU).

Having this define led to subtle bugs when using SR-IOV with a non SVP
backend like KVM. It led to crashes in the SVGA device when e.g.
shutting down a VM currently running on the VirtioGPU.

As this define prevented SR-IOV to work with KVM, and also prevented
SR-IOV with vanilla VBox, we remove it.

virtiocore: workaround lost event

The Intel DV driver does not recognize the display changed event we are
issuing. We noticed, that when setting both queue interrupt + display
event in the ISR solves this problem. It is unclear why this is the case
at the moment.

license: virtiogpu